Restructure example app with categorized navigation and image optimization (#63)
- Reorganize pages into logical categories (map/, marker/) for better structure
- Replace PNG/JPG images with WebP format for improved performance
- Add new startup page as initial screen
- Simplify sidebar navigation with categorized menu items
- Update navigation to use hierarchical page structure
- Add release build metadata and update gitignore for APK files

+++ mapconductor-core/src/main/java/com/mapconductor/settings/Settings.kt
@@ -5,8 +5,8 @@ import androidx.compose.ui.unit.dp
 
 sealed class Settings(
     val tapTolerance: Dp,
-    val markerDropAnimateDuration: Int,
-    val markerBounceAnimateDuration: Int,
+    val markerDropAnimateDuration: Long,
+    val markerBounceAnimateDuration: Long,
     val iconSize: Dp,
     val iconStroke: Dp,
     val composeEventDebounce: Long,
